# backend/VacciNationAPI/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "VacciNation Project Swagger Documentation",
    "version": "1.0.0",
    "title": "VacciNation"
  },
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/Report/staff/{staff_id}": {
      "get": {
        "tags": [
          "Report"
        ],
        "summary": "get all reporting insights for a specific staff member including name, locations, total patients prrocessed, count of adverse reactions",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "baerer token for security and autorization",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "staff_id",
            "description": "filters staff insight results by this staff_id",
            "type": "integer",
            "required": true
          },
          {
            "in": "query",
            "name": "date",
            "description": "filters staff_id insight results by this date (from this date -> present day)",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "returns information about total vaccinations given by staff member filtered by staff id and date"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Missing or invalid auth token"
          }
        }
      }
    },
    "/Report/location/{location_id}": {
      "get": {
        "tags": [
          "Report"
        ],
        "summary": "get all reporting insights for a location - includes total patients processed, total for each vaccine manufacturer, total for each sequence, total adverse reactions",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "baerer token for security and autorization",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "location_id",
            "description": "the location the staff member is getting insights for",
            "required": true,
            "type": "integer"
          },
          {
            "in": "query",
            "name": "date",
            "description": "filters vaccine insight results by this date (from this date -> present day)",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "returns information about locations filtered by date and location id"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Missing or invalid auth token"
          }
        }
      }
    },
    "/Report/reactions": {
      "get": {
        "tags": [
          "Report"
        ],
        "summary": "get all reporting insights for reactions - including date, patient name, staff name, vaccine manufacturer, brand, batch number, and reaction notes",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "baerer token for security and autorization",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "returns information about all reactions"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Missing or invalid auth token"
          }
        }
      }
    },
    "/Report/batch/{batch}": {
      "get": {
        "tags": [
          "Report"
        ],
        "summary": "get all reporting insights for a certain batch number on a certain day at a certain location - list all patients by date and location who recieved shots from a particular batch of vaccine",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "baerer token for security and autorization",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "batch",
            "type": "string",
            "required": true,
            "description": "the batch number of the vaccine you are looking to search on"
          },
          {
            "in": "query",
            "name": "date",
            "description": "the date to filter on (EXACT date to search on - no range!)",
            "type": "string",
            "required": true
          },
          {
            "in": "query",
            "name": "location_id",
            "description": "the location id to filter on",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "returns information about a batch on a specific day at a specific location"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Missing or invalid auth token"
          }
        }
      }
    },
...
  },
...
}